+++ client/presentation/src/main/java/com/andone/memorip/presentation/selectcategory/SelectCategoryScreen.kt
@@ -32,7 +32,7 @@ import com.andone.memorip.presentation.theme.LocalMemoripColors
 import com.andone.memorip.presentation.theme.LocalMemoripTypography
 import com.andone.memorip.presentation.theme.MemoripPadding.PaddingMedium
 import com.andone.memorip.presentation.theme.MemoripTheme
-import com.andone.memorip.presentation.util.dummydata.DummyData.categories
+import com.andone.memorip.presentation.util.DummyData
 
 private object Constants {
     val MAX_SELECTABLE_COUNT = 3
@@ -43,9 +43,9 @@ fun SelectCategoryScreen(
     onBackClick: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    val categories = remember{ categories }
-    val checkedSet = remember{ mutableStateSetOf<Long>() }
-    var showDialog by remember{ mutableStateOf(false) }
+    val categories = remember { DummyData.categories }
+    val checkedSet = remember { mutableStateSetOf<Long>() }
+    var showDialog by remember { mutableStateOf(false) }
 
     SelectCategoryContent(
         categories = categories,
@@ -69,7 +69,7 @@ fun SelectCategoryScreen(
             onConfirmClick = { category, color ->
                 val newCategory = Category(
                     id = (categories.maxOfOrNull { it.id } ?: -2L) + 1L,
-                    category =  category,
+                    category = category,
                     color = color
                 )
                 categories.add(newCategory)
